+++ biz/src/main/java/com/tzq/biz/common/enums/InventoryEnum.java
@@ -1,28 +1,36 @@
-package com.tzq.integration.common.enums;
+package com.tzq.biz.common.enums;
 
 import org.apache.commons.lang3.StringUtils;
 
 /**
- * 国内国际区分
+ * 舱位可销售数量枚举
+ *
  * @Author tzq24955
  * @Created on 2018/5/8
  * LY.com Inc.
  * Copyright (c) 2004-2017 All Rights Reserved.
  */
-public enum AreaTypeEnum {
-    /** 国内 */
-    DOMESTIC("10080", "国内"),
-    /** 国际 */
-    INTERNATIONAL("10081", "国际");
+public enum InventoryEnum {
 
-    /** 枚举编码 */
+    /**
+     * >=10
+     */
+    A("A", ">=10"); /** >=10 */
+
+    /**
+     * 枚举编码
+     */
     private String code;
 
-    /** 枚举描述 */
+    /**
+     * 枚举描述
+     */
     private String desc;
 
-    /** 私有构造函数 */
-    private AreaTypeEnum(String code, String desc) {
+    /**
+     * 私有构造函数
+     */
+    private InventoryEnum(String code, String desc) {
         this.code = code;
         this.desc = desc;
     }
@@ -33,11 +41,11 @@ public enum AreaTypeEnum {
      * @param code 枚举码
      * @return
      */
-    public static AreaTypeEnum getEnumByCode(String code) {
+    public static InventoryEnum getEnumByCode(String code) {
         if (StringUtils.isEmpty(code)) {
             return null;
         }
-        for (AreaTypeEnum e : AreaTypeEnum.values()) {
+        for (InventoryEnum e : InventoryEnum.values()) {
             if (code.equals(e.getCode())) {
                 return e;
             }
@@ -62,4 +70,5 @@ public enum AreaTypeEnum {
     public String getDesc() {
         return desc;
     }
+
 }
