+++ biz/src/main/java/com/tzq/biz/common/model/integration/FlightUnitVO.java
@@ -1,8 +1,8 @@
-package com.tzq.integration.common.model.integration;
+package com.tzq.biz.common.model.integration;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.tzq.integration.common.model.Money;
+import com.tzq.biz.common.model.Money;
 import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang3.StringUtils;
