+++ biz/src/main/java/com/tzq/biz/common/model/integration/StopVO.java
@@ -1,6 +1,6 @@
-package com.tzq.integration.common.model.integration;
+package com.tzq.biz.common.model.integration;
 
-import com.tzq.integration.common.model.BaseVO;
+import com.tzq.biz.common.model.BaseVO;
 
 import java.util.Date;
