+++ biz/src/main/java/com/tzq/biz/service/abstracts/AbstractBaseService.java
@@ -1,0 +1,7 @@
+package com.tzq.biz.service.abstracts;
+
+public abstract class AbstractBaseService {
+    // 统一日志记录
+    // 错误消息处理
+    // 字典配置
+}
